Move inner classes of colour picker to separate files
